refactor: refactor HatchOut DI methods
